Make all standard mounts fast
To make swift mounts a global upgrade rather than specific to the mount.
The only characters on the server with mounts currently have swift ones,
so a global override is fine for now. Once we have characters again who
only have standard mounts, we'll have to come back and make this
character-specific.